dom/presentation/tests/mochitest/test_presentation_device_info_permission.html

<!DOCTYPE HTML>
<html>
<!-- Any copyright is dedicated to the Public Domain.
   - http://creativecommons.org/publicdomain/zero/1.0/ -->
<head>
  <meta charset="utf-8">
  <title>Test for B2G Presentation Device Info API Permission</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1080474">Test for B2G Presentation Device Info API Permission</a>
<script type="application/javascript;version=1.8">

'use strict';

SimpleTest.waitForExplicitFinish();

function runTests() {
  is(navigator.mozPresentationDeviceInfo, undefined, 'navigator.mozPresentationDeviceInfo is undefined');
  SimpleTest.finish();
}

window.addEventListener('load', function() {
  SpecialPowers.pushPermissions([
    {type: 'presentation-device-manage', allow: false, context: document},
  ], function() {
    SpecialPowers.pushPrefEnv({
      'set': [
        ['dom.presentation.enabled', true],
      ]
    }, runTests);
  });
});

</script>
</pre>
</body>
</html>
